Output from code to terminal
I cannot get this code to output to a text file, I am being denied permission.  I may have made an error when setting up studio visual code or perhaps my antivirus is the problem.  I am not able to fix the issue so must submit this part of the assignment in this format.